TST split API checks from other checks
